pyproject.toml
[project]
name = "trail-condition-portal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=6.0",
    "google-genai>=1.56.0",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.2",
    "trafilatura>=2.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 80

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["trail_status/tests"]
